refactor(routing): streamline admin page structure
- Moved '/admin' route to root path '/'
- Removed deprecated admin components (AdminLayout.vue, AdminPage.vue)
- Simplified navigation structure for admin interface
