--- inventorySystem/src/pages/AdminPage.vue
@@ -1,9 +1,0 @@
-<template>
-  <q-page class="row items-center justify-evenly ">
-    <h1>Hello Admin!</h1>
-  </q-page>
-</template>
-
-<script setup lang="ts">
-
-</script>
