+++ inventorySystem/src/router/routes.ts
@@ -6,12 +6,6 @@ const routes: RouteRecordRaw[] = [
     component: () => import('layouts/MainLayout.vue'),
     children: [{ path: '', component: () => import('pages/IndexPage.vue') }],
   },
-  {
-    path: '/admin',
-    component: () => import('layouts/AdminLayout.vue'),
-    children: [{ path: '', component: () => import('pages/AdminPage.vue') }],
-  },
-
   // Always leave this as last one,
   // but you can also remove it
   {
